// front-barbershop/front-barbershop/src/features/cliente/types/cliente.types.ts
export interface BarberResponse {
  barberId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  documentNumber: string;
  commission: number;
  status: "ACTIVO" | "INACTIVO";
}